Feat: Go_channel operation using C11 Atomics and futex
The following include the function to operation as Go channel.Instead of using mutex in POSIX, we use futex to perform mutex lock and unlocked.The struct chan defines the channel for buffered and unbuffered. Theer are two types operation for unbuffered and buffered channel,which use the Atomics and futex operation.